+++ tests/conftest.py
@@ -197,7 +197,9 @@ def group_callback() -> MockTangoEventCallbackGroup:
         "kValue",
         "commandCallInfo",
         "trackTableErrors",
-        timeout=50,
+        "globalPointingModelParams",
+        "band2PointingModelParams",
+        timeout=80,
     )
     return group_callback
 
@@ -225,6 +227,10 @@ def json_factory():
 
 
 def dish_mode_callback(argin):
+    """An empty dishmode callback"""
+
+
+def pointing_model_param_callaback():
     """An empty dishmode callback"""
 
 
@@ -276,6 +282,7 @@ def cm() -> Generator[DishLNComponentManager, None, None]:
         track_table_entries=25,
         pointing_calculation_period=100,
         _update_dishmode_callback=dish_mode_callback,
+        _update_dish_pointing_model_param=(pointing_model_param_callaback),
         _update_pointingstate_callback=pointing_state_callback,
         communication_state_callback=communication_state_callback,
         component_state_callback=communication_state_callback,
@@ -317,6 +324,7 @@ def cm_without_er_lp() -> Generator[DishLNComponentManager, None, None]:
         track_table_entries=25,
         pointing_calculation_period=100,
         _update_dishmode_callback=dish_mode_callback,
+        _update_dish_pointing_model_param=(pointing_model_param_callaback),
         _event_receiver=False,
         _liveliness_probe=LivelinessProbeType.NONE,
         _update_pointingstate_callback=pointing_state_callback,
@@ -352,6 +360,7 @@ def cm_new() -> Generator[DishLNComponentManager, None, None]:
         track_table_entries=25,
         pointing_calculation_period=100,
         _update_dishmode_callback=dish_mode_callback,
+        _update_dish_pointing_model_param=(pointing_model_param_callaback),
         _update_pointingstate_callback=pointing_state_callback,
         communication_state_callback=communication_state_callback,
         component_state_callback=communication_state_callback,