+++ tests/unit/test_apply_pointing_model_command.py
@@ -6,15 +6,15 @@ from ska_tango_base.commands import ResultCode, TaskStatus
 from tests.settings import COMMAND_COMPLETION_MESSAGE
 
 
-def test_static_pm_setup_command(
+def test_apply_pointing_model_command(
     tango_context, cm, json_factory, task_callback
 ):
     """Test to check the global pointing model command
     functionality"""
     cm.get_device().update_unresponsive(False, "")
-    cm.is_staticpmsetup_allowed()
+    cm.is_ApplyPointingModel_allowed()
     global_pointing_tm_data_path = json_factory("global_pointing_model")
-    cm.static_pm_setup(
+    cm.apply_pointing_model(
         global_pointing_tm_data_path, task_callback=task_callback
     )
 
@@ -33,7 +33,7 @@ def test_static_pm_setup_command(
     )
 
 
-def test_static_pm_setup_command_with_faulty_path(
+def test_apply_pointing_model_command_with_faulty_path(
     tango_context, cm, json_factory, task_callback
 ):
     """
@@ -41,11 +41,11 @@ def test_static_pm_setup_command_with_faulty_path(
     gets detected.
     """
     cm.get_device().update_unresponsive(False, "")
-    cm.is_staticpmsetup_allowed()
+    cm.is_ApplyPointingModel_allowed()
     global_pointing_tm_model_path = json_factory("global_pointing_model")
     global_pointing_tm_model_path = json.loads(global_pointing_tm_model_path)
     global_pointing_tm_model_path["tm_data_sources"] = "abc"
-    cm.static_pm_setup(
+    cm.apply_pointing_model(
         json.dumps(global_pointing_tm_model_path), task_callback=task_callback
     )
 
@@ -72,7 +72,7 @@ def test_static_pm_setup_command_with_faulty_path(
     )
 
 
-def test_static_pm_setup_command_with_faulty_json(
+def test_apply_pointing_model_command_with_faulty_json(
     tango_context, cm, json_factory, task_callback
 ):
     """
@@ -80,12 +80,12 @@ def test_static_pm_setup_command_with_faulty_json(
     gets detected.
     """
     cm.get_device().update_unresponsive(False, "")
-    cm.is_staticpmsetup_allowed()
+    cm.is_ApplyPointingModel_allowed()
     global_pointing_tm_model_path = json_factory(
         "global_pointing_model_faulty"
     )
     global_pointing_tm_model_path = json.loads(global_pointing_tm_model_path)
-    cm.static_pm_setup(
+    cm.apply_pointing_model(
         json.dumps(global_pointing_tm_model_path), task_callback=task_callback
     )
 
@@ -105,40 +105,3 @@ def test_static_pm_setup_command_with_faulty_json(
     )
     assert "JSON Error" in call_args["call_kwargs"]["result"][1]
     assert "JSON Error" in str(call_args["call_kwargs"]["exception"])
-
-
-def test_static_pm_setup_command_with_wrong_dish_id(
-    tango_context, cm, json_factory, task_callback
-):
-    """
-    This test verifies the command gets rejected when faulty TmData path
-    gets detected.
-    """
-    cm.get_device().update_unresponsive(False, "")
-    cm.is_staticpmsetup_allowed()
-    global_pointing_tm_model_path = json_factory(
-        "global_pointing_model_ska002"
-    )
-    global_pointing_tm_model_path = json.loads(global_pointing_tm_model_path)
-    cm.static_pm_setup(
-        json.dumps(global_pointing_tm_model_path), task_callback=task_callback
-    )
-
-    task_callback.assert_against_call(
-        call_kwargs={"status": TaskStatus.QUEUED}
-    )
-    task_callback.assert_against_call(
-        call_kwargs={"status": TaskStatus.IN_PROGRESS}
-    )
-
-    call_args = task_callback.assert_against_call(
-        call_kwargs={
-            "status": TaskStatus.COMPLETED,
-            "result": mock.ANY,
-            "exception": mock.ANY,
-        }
-    )
-    assert "SKA002 is not matching" in call_args["call_kwargs"]["result"][1]
-    assert "SKA002 is not matching" in str(
-        call_args["call_kwargs"]["exception"]
-    )