+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+[0.19.0]
+********
+* Command static pointing model is renamed to ApplyPointingModel.
+
 [0.18.1]
 ********
 * Added DishLeafNode pointing tango device.